+++ CHANGELOG.md
@@ -2,7 +2,24 @@
 
 All notable changes to this project will be documented in this file
 
+## Proposed breaking changes
+ - replace naming of classes and methods using 'Workbench' to 'Bench'
+ - replace create_sample, create_container long argument lists with new XXXPost objects
+
 ## Unreleased
+
+### Added
+
+- static method Id.is_valid_id() to check if an object can be parsed as an id
+- bulk 'createContainer' methods 
+- methods 'error_results' and 'success_results' on BulkOperationResult
+- methods in Id: is_bench, is_sample
+- Classes to  define new Containers: GridContainerPost, ListContainerPost
+- Classes to define where new items are placed: ListContainerTargetLocation,
+    GridContainerTargetLocation, BenchTargetLocation, TopLevelTargetLocation
+
+### Changed
+- BREAKING: altered create_grid_container and create_list_container location arguments
 
 ## 2.2.2 2022-02-01